requirements.txt
numpy==1.21.0
pydub==0.23.1
python-telegram-bot==13.7
google-auth==1.34.0
google-cloud-texttospeech==2.5.2